Extract the execute() function
